[general.wp] Caching is (almost) done.

+++ Windows/Source/Common/CachedDataViewModel.cs
@@ -12,13 +12,16 @@ namespace PocketCampus.Common
     /// <typeparam name="TData">The type of the cached data.</typeparam>
     public abstract class CachedDataViewModel<TArg, TData> : DataViewModel<TArg>
     {
-        private const int DefaultId = 0;
+        private const long DefaultId = 0;
         private static readonly DateTime DefaultExpirationDate = DateTime.MaxValue;
 
         private readonly IDataCache _cache;
 
         private CacheStatus _cacheStatus;
 
+        /// <summary>
+        /// Gets the cache status of the ViewModel.
+        /// </summary>
         public CacheStatus CacheStatus
         {
             get { return _cacheStatus; }
@@ -36,11 +39,29 @@ namespace PocketCampus.Common
         }
 
 
+        /// <summary>
+        /// Gets refreshed data that can be cached, or explicit instructions to not cache it.
+        /// </summary>
+        /// <param name="force">A value indicating whether to force a refresh or not.</param>
+        /// <param name="token">The cancellation token for the refresh request.</param>
+        /// <returns>A <see cref="CachedTask" /> instance indicating what to get and whether to cache it.</returns>
         protected abstract CachedTask<TData> GetData( bool force, CancellationToken token );
 
+        /// <summary>
+        /// Handles data, either cached or live.
+        /// </summary>
+        /// <param name="data">The data.</param>
+        /// <param name="token">The cancellation token for the refresh request.</param>
+        /// <returns>True if the data should be cached, false otherwise.</returns>
         protected abstract bool HandleData( TData data, CancellationToken token );
 
 
+        /// <summary>
+        /// Asynchronously refreshes the data using cached data both as a placeholder and as a fallback.
+        /// </summary>
+        /// <param name="token">The cancellation token for the refresh request.</param>
+        /// <param name="force">A value indicating whether to force a refresh or not.</param>
+        /// <returns>A Task that will complete when the refresh operation is done.</returns>
         protected override sealed async Task RefreshAsync( CancellationToken token, bool force )
         {
             var cachedData = GetData( force, token );
